Report no running Lustre subsystem as success
When the metrics for {mdt,ost,osc,router} are requested, return
success if lustre is not running, or the subsystem is not running.

This fixes a bug introduced in:
* f984897 handle lustre version file containing bare version

Define a message string for these cases, respectively:
	no_running_mdt
	no_running_ost
	no_running_osc
	no_running_lnet_router

When cerebrod is started upon node boot, these services are typically
not running, but they may be started later.  From cerebro's perspective,
this is not an error, but it should be reported so consumers can
monitor the state of the system.

Signed-off-by: Olaf Faaland <faaland1@llnl.gov>