replace hostname with uname -n in t10.exp
Test t10-metric-strings files with expected output, t10.exp,
are evaled before comparison.  The metric strings they are
comparing are produced by lmtmetric, and include the hostname.
The eval puts the hostname into the expected string so it does
not result in an irrelevant difference and cause the comparison
to fail.

The 2.8.10 based tests had `$(hostname)` in the expected files,
which works fine on TOSS systems but not under our build system.

Replace with `$(uname -n)` which does work under koji.

Signed-off-by: Olaf Faaland <faaland1@llnl.gov>